tests: extract extras and stats tests module
