+++ tests/test_stats_extra.h
@@ -1,0 +1,8 @@
+#ifndef TEST_STATS_EXTRA_H
+#define TEST_STATS_EXTRA_H
+
+void test_extra(void);
+void test_memstats(void);
+void test_stats(void);
+
+#endif
